add: refactor function and create packages
