+++ .gitignore
@@ -7,3 +7,4 @@ link_nodes.csv
 nodes.csv
 out_nodes.csv
 prompt.txt
+.pytest_cache
