--- link_nodes.csv
@@ -1,40 +1,0 @@
-Origem da linha,Destino da linha
-14,14
-14,14
-14,14
-8,8
-15,15
-20,20
-16,20
-16,18
-18,23
-4,6
-4,7
-6,5
-5,9
-5,8
-9,8
-9,10
-9,11
-7,5
-14,8
-8,13
-8,12
-14,15
-14,16
-16,17
-16,19
-19,21
-20,24
-20,22
-20,25
-28,27
-28,29
-27,30
-26,31
-4,26
-26,28
-27,32
-28,33
-26,34
-31,35
