+++ main.py
@@ -1,28 +1,29 @@
-from load_csv import LoadCsv
-from convert_csv import CSVConverter
-from parent_nodes import ParentNodes
-from link_nodes import LinkNodes
-from menu import Menu
-from anytree import RenderTree, AsciiStyle
+from utils_csv.load_csv import LoadCsv
+from utils_csv.convert_csv import CSVConverter
+
+from builder_tree.parent_nodes import ParentNodes
+from builder_tree.link_nodes import LinkNodes
+
+from menus.menu import Menu
 
 # load the CSV file
-apex = LoadCsv('data_nodes.csv', 'nodes.csv', 'out_nodes.csv', 'temp_response.csv')
-apex.create_nodes()
-apex.create_out_nodes()
-apex.create_temp_res_nodes()
+apex = LoadCsv('nodes_csv/data_nodes.csv')
+apex.create_load_csv('nodes_csv/nodes.csv', 'Id', 'Área de texto 1')
+apex.create_load_csv('nodes_csv/out_nodes.csv', 'Origem da linha', 'Destino da linha')
+apex.create_load_csv('nodes_csv/temp_response.csv', 'Id', 'r')
 
 # convert items in the CSV file
-converter = CSVConverter('out_nodes.csv', 'link_nodes.csv')
+converter = CSVConverter('nodes_csv/out_nodes.csv', 'nodes_csv/link_nodes.csv')
 converter.convert()
 
-# create the nodes
-main_apex = ParentNodes('nodes.csv', 'temp_response.csv')
+# create nodes
+main_apex = ParentNodes('nodes_csv/nodes.csv', 'nodes_csv/temp_response.csv')
 node_menu = main_apex.create_nodes()
 main_apex.display_nodes(node_menu)
 node_menu_resp = main_apex.add_answer_nodes(node_menu)
 
 # linking nodes
-child_nodes = LinkNodes('link_nodes.csv')
+child_nodes = LinkNodes('nodes_csv/link_nodes.csv')
 nodes = child_nodes.create_link_nodes(node_menu_resp)
