--- nodes.csv
@@ -1,37 +1,0 @@
-Id,Área de texto 1
-1,Peterinvest chatbot
-2,Mapa mental com grupos coloridos
-3,Apresentar as perguntas do nó destino. Se houver menos de 3 perguntas fazer a busca recursiva da raiz.
-4,Chatbot
-5,Conector
-6,Ela distribui dividendos?
-7,O que é Dividend Yield?
-8,"O que é mais importante dividendos ou lucro?
-[R]"
-9,Porque as pessoas buscam Diviend Yield acima de 6%?
-10,Porque empresas que pagam muito DY tendem a crescer menos do que as que pagam mais?
-11,Vale a pena comprar ações que pagam menos de 6% de Dividend Yield?
-12,Porque ver o histórico de lucros é importante?
-13,Pode me mostrar uma comparação entre cotação e lucros?
-14,Por que normalmente o setor elétrico paga bons dividendos?
-15,Porque o setor de transmissão é o mais previsível?
-16,"Qual a diferença entre transmissão, geração e distribuição?"
-17,Pode me mostrar um vídeo de um minuto de como funcionam esses setores? (mostra um vídeo)
-18,Como funciona o setor de distribuição?
-19,Como funciona o setor de geração?
-20,Como funciona o setor de transmissão?
-21,Porque a receita da geradora é variável?
-22,Como uma empresa consegue um contrato para fazer uma linha de transmissão?
-23,Por que a distribuidora é o mais difícil setor?
-24,O que é ANEEL?
-25,Porque o setor de transmissão é o mais previsível?
-26,O que é P/L?
-27,Me mostre o gráfico histórico do P/L.
-28,Qual o P/L do ativo? (contexto de ativo tudo abaixo)
-29,"Qual a média do P/L histórico ajustado, comparado com pares internacionais e nacionais? (Premium)"
-30,Qual o P/L do setor?
-31,Qual a média do P/L da bolsa?
-32,Para quê serve o histórico do P/L?
-33,O que é média histórica justada e para que serve? 
-34,Porque existem empresas com P/L muito alto e muito baixo?
-35,Qual a média do P/L da bolsa dos últimos 20 anos? Me apresente um gráfico.
