--- out_nodes.csv
@@ -1,40 +1,0 @@
-Origem da linha,Destino da linha
-14.0,14.0
-14.0,14.0
-14.0,14.0
-8.0,8.0
-15.0,15.0
-20.0,20.0
-16.0,20.0
-16.0,18.0
-18.0,23.0
-4.0,6.0
-4.0,7.0
-6.0,5.0
-5.0,9.0
-5.0,8.0
-9.0,8.0
-9.0,10.0
-9.0,11.0
-7.0,5.0
-14.0,8.0
-8.0,13.0
-8.0,12.0
-14.0,15.0
-14.0,16.0
-16.0,17.0
-16.0,19.0
-19.0,21.0
-20.0,24.0
-20.0,22.0
-20.0,25.0
-28.0,27.0
-28.0,29.0
-27.0,30.0
-26.0,31.0
-4.0,26.0
-26.0,28.0
-27.0,32.0
-28.0,33.0
-26.0,34.0
-31.0,35.0
