--- prompt.txt
@@ -1,81 +1,0 @@
-could you create nodes from csv file, in python anytree, here is an example if csv file:
-Id,Área de texto 1
-1,Peterinvest chatbot
-2,Mapa mental com grupos coloridos
-3,Apresentar as perguntas do nó destino. Se houver menos de 3 perguntas fazer a busca recursiva da raiz.
-4,Chatbot
-5,Conector
-6,Ela distribui dividendos?
-7,O que é Dividend Yield?
-8,O que é mais importante dividendos ou lucro?
-9,Porque as pessoas buscam Diviend Yield acima de 6%?
-10,Porque empresas que pagam muito DY tendem a crescer menos do que as que pagam mais?
-11,Vale a pena comprar ações que pagam menos de 6% de Dividend Yield?
-12,Porque ver o histórico de lucros é importante?
-13,Pode me mostrar uma comparação entre cotação e lucros?
-14,Por que normalmente o setor elétrico paga bons dividendos?
-15,Porque o setor de transmissão é o mais previsível?
-16,"Qual a diferença entre transmissão, geração e distribuição?"
-17,Pode me mostrar um vídeo de um minuto de como funcionam esses setores? (mostra um vídeo)
-18,Como funciona o setor de distribuição?
-19,Como funciona o setor de geração?
-20,Como funciona o setor de transmissão?
-21,Porque a receita da geradora é variável?
-22,Como uma empresa consegue um contrato para fazer uma linha de transmissão?
-23,Por que a distribuidora é o mais difícil setor?
-24,O que é ANEEL?
-25,Porque o setor de transmissão é o mais previsível?
-and use this another example of csv file to create links between nodes:
-Origem da linha,Destino da linha
-14,14
-14,14
-14,14
-8,8
-15,15
-20,20
-16,20
-16,18
-18,23
-4,6
-4,7
-6,5
-5,9
-5,8
-9,8
-9,10
-9,11
-7,5
-14,8
-8,13
-8,12
-14,15
-14,16
-16,17
-16,19
-19,21
-20,24
-20,22
-20,25
-the 'Origem da linha' is a node and the 'Destino da linha' is the node that the 'Origem da linha' is connected to.
-if exist a node that refers to itself, like in the example, that node point to null, write a test.
-================================================================================================
-blz. é só ir voltando e buscando pra frente as perguntas até somarem 3.O
-usuário pode dizer "[Não sugira mais essa pergunta]"
-================================================================================================
-Aqui está um exemplo genérico de uma função de pesquisa recursiva em Python:
-
-python
-Copy code
-def pesquisa_recursiva(nodo, valor):
-    if nodo.valor == valor:
-        return nodo
-
-    for filho in nodo.filhos:
-        resultado = pesquisa_recursiva(filho, valor)
-        if resultado is not None:
-            return resultado
-
-    return None
-Essa função pressupõe que o objeto nodo possui um atributo valor e uma lista de filhos. Note que esse exemplo é apenas uma base genérica e pode não funcionar diretamente com a estrutura de árvore que você está usando. Será necessário adaptar a função de acordo com sua implementação específica.
-
-Você também pode considerar o uso de bibliotecas Python, como a anytree [3], que pode facilitar a criação e manipulação de árvores genéricas. No entanto, a resposta para realizar a pesquisa recursiva em uma árvore genérica usando a biblioteca anytree não está disponível nos resultados da pesquisa fornecidos.
